Refactor itemMediaInfo
