+++ src/index.d.ts
@@ -2,3 +2,8 @@ declare module '*.png' {
    const value: any;
    export = value;
 }
+
+declare module '*.html' {
+   const content: string;
+   export default content;
+}
